# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AfriSense API",
    "version": "1.0.0",
    "description": "Documentation OpenAPI minimale pour les endpoints d'authentification et trackers"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Créer un utilisateur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "phone": { "type": "string" },
                  "password": { "type": "string" },
                  "role": { "type": "string" }
                },
                "required": ["fullName","email","password","role"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Utilisateur créé" },
          "400": { "description": "Erreur de validation" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Connexion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email","password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Connexion réussie" },
          "401": { "description": "Identifiants invalides" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Profil de l'utilisateur connecté",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Profil retourné" },
          "401": { "description": "Non authentifié" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}